pyproject.toml
[tool.poetry]
name = "tsukuba-paypay-campain"
version = "0.1.0"
description = ""
authors = ["taki-d <mail@orangepla.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.1.2"
tabula-py = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
